backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
requests==2.32.3
python-dateutil==2.9.0 